Refactor some code back
